move recentlyPlayed data to PlayerContext in Player.js
