+++ jukeos/src/components/Home.js
@@ -119,14 +119,21 @@ const ScrollWheel = ({ items }) => {
 
 const Home = () => {
   const { accessToken, invalidateAccess } = useContext(SpotifyAuthContext);
-  const { track, paused, playUri, togglePlay,nextTrack,prevTrack } = useContext(PlayerContext);
+  const {
+    track,
+    paused,
+    playUri,
+    togglePlay,
+    nextTrack,
+    prevTrack,
+    recentlyPlayed,
+    recentlyPlayedError,
+    isLoadingRecent
+  } = useContext(PlayerContext);
   const [currentTrack, setCurrentTrack] = useState({
     progress: 0,
     duration: 1
   });
-  const [recentlyPlayed, setRecentlyPlayed] = useState([]);
-  const [recentlyPlayedError, setRecentlyPlayedError] = useState(null);
-  const [isLoadingRecent, setIsLoadingRecent] = useState(true);
 
   const formatTime = (seconds) => {
     const minutes = Math.floor(seconds / 60);
@@ -158,72 +165,6 @@ const Home = () => {
       console.error('Failed to seek:', error);
     }
   };
-
-  /**
-   * Backend Requirements for Recently Played Tracks:
-   * 
-   * This frontend code calls Spotify's /me/player/recently-played endpoint which requires:
-   * 1. A valid Spotify access token in the Authorization header
-   * 2. Returns up to 20 most recently played tracks
-   * 
-   * Backend Team Needs to:
-   * - Implement token refresh mechanism to ensure valid access tokens
-   * - Consider caching recently played tracks to reduce API calls
-   * - Handle rate limiting (Spotify allows 1 request/sec)
-   * - Implement error handling for expired/invalid tokens
-   * - Consider implementing a proxy endpoint to hide Spotify credentials
-   *   Example: /api/recently-played instead of calling Spotify directly
-   * 
-   * Relevant Documentation: https://developer.spotify.com/documentation/web-api/reference/get-recently-played
-   */
-  const fetchRecentlyPlayed = () => {
-    if (accessToken) {
-      setIsLoadingRecent(true);
-
-      performFetch("https://api.spotify.com/v1/me/player/recently-played", { limit: 10 }, accessToken, invalidateAccess)
-        .then((response) => {
-          console.log("Successfully fetched recently played:", response);
-
-          if (response && response.items) {
-            // Transform the data to match our UI needs
-            const transformedTracks = response.items
-              .filter((item) => item && item.track && item.track.album)
-              .map((item) => ({
-                id: item.track.id,
-                title: item.track.name,
-                artist: item.track.artists[0].name,
-                imageUrl: item.track.album.images[0]?.url || defaultAlbumArt,
-                playedAt: new Date(item.played_at),
-                // Add any additional track data you need
-                albumName: item.track.album.name,
-                duration: item.track.duration_ms,
-                uri: item.track.uri
-              }))
-              .sort((a, b) => b.playedAt.getTime() - a.playedAt.getTime())
-
-            setRecentlyPlayed(transformedTracks);
-          }
-        })
-        .catch((error) => {
-          console.error("Failed to fetch recently played:", error);
-          setRecentlyPlayedError(error);
-        })
-        .finally(() => {
-          setIsLoadingRecent(false);
-        });
-    }
-  };
-
-  useEffect(() => {
-    if (accessToken) {
-      fetchRecentlyPlayed();
-
-      // Optional: Set up polling to keep recently played list updated
-      const pollInterval = setInterval(fetchRecentlyPlayed, 30000); // 30 seconds
-
-      return () => clearInterval(pollInterval);
-    }
-  }, [accessToken]);
 
   useEffect(() => {
     if (!accessToken) return;
