+++ jukeos/src/components/Player.js
@@ -1,5 +1,6 @@
 import {useState, createContext, useContext, useEffect, useCallback} from "react";
-import { SpotifyAuthContext, performPut } from "../contexts/spotify";
+import { SpotifyAuthContext, performFetch, performPut } from "../contexts/spotify";
+import defaultAlbumArt from '../assets/default-art-placeholder.svg';
 
 
 const Player = ({ children }) => {
@@ -11,6 +12,9 @@ const Player = ({ children }) => {
   const [active, setActive] = useState(false);
   const [track, setTrack] = useState(null);
   const [paused, setPaused] = useState(true);
+  const [recentlyPlayed, setRecentlyPlayed] = useState([]);
+  const [recentlyPlayedError, setRecentlyPlayedError] = useState(null);
+  const [isLoadingRecent, setIsLoadingRecent] = useState(true);
 
   //TODO Refactor a lot of this listener logic
   useEffect(() => {
@@ -61,6 +65,73 @@ const Player = ({ children }) => {
     }
   }, [accessToken, playbackReady, player]);
 
+  /**
+   * Backend Requirements for Recently Played Tracks:
+   * 
+   * This frontend code calls Spotify's /me/player/recently-played endpoint which requires:
+   * 1. A valid Spotify access token in the Authorization header
+   * 2. Returns up to 20 most recently played tracks
+   * 
+   * Backend Team Needs to:
+   * - Implement token refresh mechanism to ensure valid access tokens
+   * - Consider caching recently played tracks to reduce API calls
+   * - Handle rate limiting (Spotify allows 1 request/sec)
+   * - Implement error handling for expired/invalid tokens
+   * - Consider implementing a proxy endpoint to hide Spotify credentials
+   *   Example: /api/recently-played instead of calling Spotify directly
+   * 
+   * Relevant Documentation: https://developer.spotify.com/documentation/web-api/reference/get-recently-played
+   */
+  const fetchRecentlyPlayed = () => {
+    if (accessToken) {
+      setIsLoadingRecent(true);
+
+      performFetch("https://api.spotify.com/v1/me/player/recently-played", { limit: 10 }, accessToken, invalidateAccess)
+        .then((response) => {
+          console.log("Successfully fetched recently played:", response);
+
+          if (response && response.items) {
+            // Transform the data to match our UI needs
+            const transformedTracks = response.items
+              .filter((item) => item && item.track && item.track.album)
+              .map((item) => ({
+                id: item.track.id,
+                title: item.track.name,
+                artist: item.track.artists[0].name,
+                imageUrl: item.track.album.images[0]?.url || defaultAlbumArt,
+                playedAt: new Date(item.played_at),
+                // Add any additional track data you need
+                albumName: item.track.album.name,
+                duration: item.track.duration_ms,
+                uri: item.track.uri
+              }))
+              .sort((a, b) => b.playedAt.getTime() - a.playedAt.getTime())
+
+            setRecentlyPlayed(transformedTracks);
+          }
+        })
+        .catch((error) => {
+          console.log(setRecentlyPlayed);
+          console.log(typeof(setRecentlyPlayed));
+          console.error("Failed to fetch recently played:", error);
+          setRecentlyPlayedError(error);
+        })
+        .finally(() => {
+          setIsLoadingRecent(false);
+        });
+    }
+  };
+
+  useEffect(() => {
+    if (accessToken) {
+      fetchRecentlyPlayed();
+
+      // Optional: Set up polling to keep recently played list updated
+      const pollInterval = setInterval(fetchRecentlyPlayed, 30000); // 30 seconds
+
+      return () => clearInterval(pollInterval);
+    }
+  }, [accessToken]);
 
   //Use effect to keep track of the current Track.
   useEffect(() => {
@@ -70,7 +141,8 @@ const Player = ({ children }) => {
       player.getCurrentState().then((state) => {
         if (state && state.track_window.current_track) {
           setTrack(state.track_window.current_track);
-          console.log("Setting Track: " + JSON.stringify(track));
+          console.log("Setting Track: " + JSON.stringify(state.track_window.current_track));
+          // console.log(recentlyPlayed);
           setPaused(state.paused);
           setActive(!!state);
         }
@@ -148,7 +220,10 @@ const Player = ({ children }) => {
         togglePlay,
         playUri,
         nextTrack,
-        prevTrack
+        prevTrack,
+        recentlyPlayed,
+        recentlyPlayedError,
+        isLoadingRecent
       }
     }>
       { children }
@@ -167,5 +242,8 @@ export const PlayerContext = createContext({
   togglePlay: () => {},
   playUri: () => {},
   nextTrack: () => {},
-  prevTrack: () => {}
+  prevTrack: () => {},
+  recentlyPlayed: [],
+  recentlyPlayedError: null,
+  isLoadingRecent: true
 });
